Extract ItemsSubscription from Subscription
Subscription is now an abstract class, and ItemsSubscription is a
concrete class of Subscription.
It is responsible for finding new items in the feed and provides
unreadItems/readItems/lastFoundItems features.

Also it adds SubscriptionFactory that creates an instance of
Subscription subclass depending on `type` value.
And Subscriber#itemsSubscription getter returns a list of
ItemsSubscriptions only.