chore(package): Prepeare for publishing as NPM module
To introduce semantic-release, it is required to publish a module to NPM for automatic versioning.
We will publish an extension as `@io-monad/novelous-extension`, using namespace but a public module.
To publish to NPM, we update package.json by separating `dependencies` from `devDependencies` and
add `files` field not to publish unnecessary resource files such as images and fonts.